ouvir-para-ver-backend/requirements.txt
gradio
torch
transformers
pillow
sentencepiece
